Updates for handling breaking changes introduced by OpenZeppelin.
OZ released v4.6.0 on 26 April 2022 and it has breaking changes [1].

This PR
- Updates GovernorContract, GovernanceTokenWrapper.

Note that as of today the Open Zeppelin Wizard's auto-generated
code does not reflect the breaking changes.

TESTS: ran `yarn hardhat test --deploy-fixture`. The flag wasn't necessary but added it
as some people may need it depending on their config.

[1] https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/CHANGELOG.md#breaking-changes